requirements.txt
# ComfyUI API
websocket-client==1.6.4
requests==2.31.0

# Image Processing
Pillow==10.1.0

# Twitter API
tweepy==4.14.0

# Solana & NFT
solders==0.18.1
solana==0.30.2
anchorpy==0.18.0

# Scheduling
schedule==1.2.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pytz==2023.3

# Web Automation
playwright==1.40.0
